# app/static/swagger.yaml
swagger: "2.0"
info:
  title: Blog API
  description: >
    This is the backend to a blog app that allows users to register, create blogs, comment and like blogs...
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

paths:
  /users/login:
    post:
      tags:
        - User
      summary: Login a user
      description: >
        Endpoint to login a user, all fields are required. Takes in an email and password and returns a jwt with users id, email, username, and name.
      parameters:
        - in: body
          name: user
          description: User login data
          required: true
          schema:
            $ref: "#/definitions/UserLogin"
      responses:
        200:
          description: Login successful
          schema:
            $ref: "#/definitions/LoginResponse"

        400:
          description: Missing required field
          schema:
            type: object
            example:
              error: "Email and password are required"

        401:
          description: Invalid email or password
          schema:
            type: object
            example:
              error: "Invalid email or password"

        500:
          description: Internal server error
          schema:
            type: object
            example:
              error: "Internal server error"
              details: "error details here..."

  /users:
    post:
      tags:
        - User
      summary: Register a user
      description: >
        Endpoint accepts 'name', 'username', 'email', and 'password', all are required. Username must be between 3-150 characters. Email must be in email format (i.e 'test@test.com'). Name, username, email, and password are all strings.
      parameters:
        - in: body
          name: user data
          description: User registration data
          required: true
          schema:
            $ref: "#/definitions/UserRegistration"
      responses:
        201:
          description: User created successfully
          schema:
            $ref: "#/definitions/UserResponse"

        400:
          description: Validation/Integrity error
          schema:
            type: object
            example:
              error: "Not a valid email"

        409:
          description: Email already registered
          schema:
            type: object
            example:
              error: "Email already registered"

        500:
          description: Internal server error
          schema:
            type: object
            example:
              error: "Internal server error"
              details: "error details here..."

  /users/me:
    patch:
      tags:
        - User
      summary: Update Currently logged in user
      description: >
        Authenticated endpoint, user must be logged in with a valid JWT. To update a user, only enter the data that needs to be changed (this is a patch request not a put request). ***Do NOT use this route to update password, please use /users/updatePassword.***
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: user data
          description: update user data
          required: true
          schema:
            type: object
            properties:
              username:
                type: string
                example: "jdoe2020"
              email:
                type: string
                format: email
                example: "jdoe2020@example.com"
      responses:
        200:
          description: Successfully updated user
          schema:
            $ref: "#/definitions/UpdateUserResponse"

        400:
          description: >
            Validation error (i.e you tried update your password or entered an invalid request)
          schema:
            type: object
            example:
              error: "Do not use this route to update password. Use /users/changePassword"

        401:
          description: >
            Authentication failed. Possible reasons include:
              - Missing token
              - Invalid token
              - Expired token
          schema:
            type: object
            example:
              error: "Token is missing"

        500:
          description: Internal server error
          schema:
            type: object
            example:
              error: "Internal server error"
              details: "error details here..."

  /blogs:
    post:
      tags:
        - Blog
      summary: Create a blog
      description: >
        This is an authenticated route, only title and body are required. Other fields are not to be updated directly.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: blog data
          description: Data to create a new blog entry
          required: true
          schema:
            $ref: "#/definitions/CreateBlog"
      responses:
        201:
          description: Blog created successfully
          schema:
            $ref: "#/definitions/CreateBlogResponse"

        400:
          description: Validation / Integrity error (i.e you entered invalid data as a request)
          schema:
            type: object
            example:
              error: "validation error here..."

        401:
          description: >
            Authentication failed. Possible reasons include:
              - Missing token
              - Invalid token
              - Expired token
          schema:
            type: object
            example:
              error: "Token is missing"

        404:
          description: No user found (i.e you deleted your account and tried creating a blog post)
          schema:
            type: object
            example:
              message: "No user found"

        500:
          description: Internal server error
          schema:
            type: object
            example:
              error: "Internal server error"
              details: "error details here..."

    get:
      tags:
        - Blog
      summary: Get all blogs
      description: Endpoint to retrieve a list of all blogs in the database.
      responses:
        200:
          description: A list of blogs
          schema:
            type: array
            items:
              $ref: "#/definitions/blogs"
            examples:
              application/json:
                - id: 1
                  author_id: 1
                  title: "Test Blog"
                  body: "blah blah blah"
                  created_at: "2025-07-15T08:01:30"
                  is_archived: false
                  updated_at: null
                - id: 2
                  author_id: 1
                  title: "Test Blog Two"
                  body: "blah blah blah"
                  created_at: "2025-07-17T00:26:08"
                  is_archived: false
                  updated_at: null

        500:
          description: Internal server error
          schema:
            type: object
            example:
              error: "Internal server error"
              details: "error details here..."

  /blogs/{blog_id}:
    get:
      tags:
        - Blog
      summary: Get a blog by id
      description: Takes on parameter (blog_id) and returns a single blog.
      parameters:
        - in: path
          name: blog_id
          required: true
          type: integer
          description: ID of the blog to retrieve
      responses:
        200:
          description: blog retrieved successfully
          schema:
            $ref: "#/definitions/CreateBlogResponse"

        404:
          description: Blog not found
          schema:
            type: object
            example:
              error: "No blog found"

        500:
          description: Internal server error
          schema:
            type: object
            example:
              error: "Internal server error"
              details: "error details here..."

definitions:
  UserRegistration:
    type: object
    required:
      - name
      - username
      - email
      - password
    properties:
      name:
        type: strings
        example: "John Doe"
      username:
        type: string
        example: "jdoe20"
      email:
        type: string
        format: email
        example: "jdoe@example.com"
      password:
        type: string
        format: password
        example: "test123"

  UserResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      username:
        type: string
        example: "jdoe20"
      email:
        type: string
        example: "jdoe@example.com"
      created_at:
        type: string
        example: "2025-07-15T08:01:20"

  UserLogin:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
        example: "jdoe@example.com"
      password:
        type: string
        format: password
        example: "test123"

  LoginResponse:
    type: object
    properties:
      status:
        type: string
        example: "success"
      message:
        type: string
        example: "Login successful"
      token:
        type: string
        example: "al;knv2knv[poikvnvknawnv..."
      user:
        type: object
        properties:
          id:
            type: string
            example: 1
          name:
            type: string
            example: "John Doe"
          username:
            type: string
            example: "jdoe20"

  UpdateUserResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      username:
        type: string
        example: "jdoe2020"
      email:
        type: string
        format: email
        example: "jdoe2020@example.com"
      created_at:
        type: string
        example: "2025-07-15T08:01:20"

  CreateBlog:
    type: object
    required:
      - title
      - body
    properties:
      title:
        type: string
        example: "Example Blog"
      body:
        type: string
        example: "Blah blah blah"

  CreateBlogResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      author_id:
        type: integer
        example: 1
      title:
        type: string
        example: "Example Blog"
      body:
        type: string
        example: "Blah blah blah"
      created_at:
        type: string
        example: "2025-07-15T08:01:30"
      updated_at:
        type: string
        example: null
      is_archived:
        type: boolean
        example: false